Add files via upload
This was made to test how using both a cutoff for the importance of the mean decrease accuracy (MDA), frequency of features showing up in each Random Forest run, and the number of runs for Random Forest. The idea was to see if we could get more accurate results with our Random Forest, since if we reversed the order (for example, comparing G1 to G3 vs G3 to G1) most of our features matched. But, we were seeing different "outliers". We found the most consistent minimum (to avoid overfitting) was to have MDA cutoff > 1.5, freq >70%, and trained 100 times.